Немного кода к ДЗ № 2
