--- home_work_2/lesson_4/lesson_4/task_1-3/data/schema_list_all_breeds.json
@@ -1,2 +1,0 @@
-{
-}
