+++ home_work_2/lesson_4/task_1-3/data/schema_list_all_breeds.json
@@ -1,0 +1,25 @@
+{
+    "type": "object",
+    "required": [
+        "message",
+        "status"
+    ],
+    "properties": {
+        "message": {
+            "type": "object",
+            "additionalProperties": false,
+            "patternProperties": {
+                "^.*$": {
+                    "type": "array",
+                    "additionalItems": true,
+                    "items": {
+                        "type": "string"
+                    }
+                }
+            }
+        },
+        "status": {
+            "type": "string"
+        }
+    }
+}
