+++ home_work_2/lesson_4/task_1-3/data/schema_random_image.json
@@ -1,0 +1,15 @@
+{
+    "type": "object",
+    "required": [
+        "message",
+        "status"
+    ],
+    "properties": {
+        "message": {
+            "pattern": ".*\\.jpg$"
+        },
+        "status": {
+            "type": "string"
+        }
+    }
+}
